Simplifies font testing, adds filtering to runner
DrawString tests have been rearranged so that their output is in a
single 'DrawString' directory, rather than separate directories for each
test.

CommandLineInterface now implements regular expression-based filtering
of tests.  Use -i=/regex/ or -i=PlainText to include tests.  Use
-x=/regex/ or -x=PlainText to exclude tests.  Multiple -i and -x flags
may be chained together and are applied in order to constrict the set
of tests that are run.